# server/src/main/resources/openapi/meta-storage-1.0.yaml
openapi: 3.0.0
info:
  title: Meta Storage
  version: v1
paths:
  # Configuration
  /meta-storage/config/matchkeys:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
    post:
      description: Create match key
      operationId: postConfigMatchKey
      requestBody:
        content:
          application/json:
            schema:
              $ref: schemas/matchKey.json
      responses:
        "201":
          description: Match key configuration created
          content:
            application/json:
              schema:
                $ref: schemas/matchKey.json
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
    get:
      description: Get match key configurations
      operationId: getConfigMatchKeys
      parameters:
        - $ref: parameters/limit.yaml
        - $ref: parameters/query.yaml
        - $ref: parameters/offset.yaml
      responses:
        "200":
          description: list of match keys.
          content:
            application/json:
              schema:
                $ref: schemas/matchKeys.json
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
  /meta-storage/config/matchkeys/{id}:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: id
        required: true
        description: match key identifier
        schema:
          type: string
    get:
      description: Get match key configuration
      operationId: getConfigMatchKey
      responses:
        "200":
          description: Match key configuration returned
          content:
            application/json:
              schema:
                $ref: schemas/matchKey.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    delete:
      description: Delete match key configuration
      operationId: deleteConfigMatchKey
      responses:
        "204":
          description: match key configuration deleted
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    put:
      description: Update match key configuration.
      operationId: putConfigMatchKey
      requestBody:
        content:
          application/json:
            schema:
              $ref: schemas/matchKey.json
      responses:
        "204":
          description: Match Key configuration updated
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  /meta-storage/config/matchkeys/{id}/initialize:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: id
        required: true
        description: match key identifier
        schema:
          type: string
    put:
      description: Recalculate match key across all records.
      operationId: initializeMatchKey
      responses:
        "200":
          description: match key statistics.
          content:
            application/json:
              schema:
                $ref: schemas/matchKeyStats.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  /meta-storage/config/matchkeys/{id}/stats:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: id
        required: true
        description: match key identifier
        schema:
          type: string
    get:
      description: Get statistics for match key configuration
      operationId: statsMatchKey
      responses:
        "200":
          description: match key statistics.
          content:
            application/json:
              schema:
                $ref: schemas/matchKeyStats.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  # Sources
  /meta-storage/sources:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
    post:
      description: Create source.
      operationId: postSource
      requestBody:
        content:
          application/json:
            schema:
              $ref: schemas/source.json
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: schemas/source.json
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
    get:
      description: Get sources.
      operationId: getSources
      parameters:
        - $ref: parameters/limit.yaml
        - $ref: parameters/query.yaml
        - $ref: parameters/offset.yaml
      responses:
        "200":
          description: list of sources.
          content:
            application/json:
              schema:
                $ref: schemas/sources.json
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
  /meta-storage/config/sources/{id}:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: id
        required: true
        description: source identifier
        schema:
          type: string
          format: uuid
    get:
      description: Get source.
      operationId: getSource
      responses:
        "200":
          description: get match key.
          content:
            application/json:
              schema:
                $ref: schemas/matchKey.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    delete:
      description: Delete source.
      operationId: deleteSource
      responses:
        "204":
          description: source deleted.
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  # Ingest
  /meta-storage/ingest-jobs:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
    post:
      description: Create ingest job
      operationId: postIngestJob
      requestBody:
        content:
          application/json:
            schema:
              $ref: schemas/ingestJobRequest.json
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: schemas/ingestJobResponse.json
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
  /meta-storage/ingest-jobs/{id}:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: id
        required: true
        description: ingest job identifier.
        schema:
          type: string
          format: uuid
    put:
      description: Put records for job.
      operationId: ingestJobRecord
      requestBody:
        content:
          application/json:
            schema:
              $ref: schemas/ingestRecordChunk.json
      responses:
        "200":
          description: Records accepted.
          content:
            application/json:
              schema:
                $ref: schemas/ingestRecordResponse.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    get:
      description: Get ingest job information.
      operationId: ingestJobInfo
      responses:
        "200":
          description: ingest job found OK.
          content:
            application/json:
              schema:
                $ref: schemas/ingestJobInfo.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    delete:
      description: Finish ingest job with either rollback of commit.
      operationId: ingestJobFinish
      parameters:
        - in: query
          name: commit
          required: false
          description: whether to commit
          schema:
            type: boolean
            default: false
      responses:
        "204":
          description: Job deleted.
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  # Global Records retrieval and simple ingest
  /meta-storage/records:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
    get:
      description: Get records that satisfy CQL query with fields localId, sourceId, globalId.
      operationId: getGlobalRecords
      parameters:
        - $ref: parameters/limit.yaml
        - $ref: parameters/query.yaml
        - $ref: parameters/offset.yaml
      responses:
        "200":
          description: globalindex record(s) found
          content:
            application/json:
              schema:
                $ref: schemas/globalRecords.json
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
    put:
      description: Create or update records.
      operationId: putGlobalRecords
      requestBody:
        content:
          application/json:
            schema:
              $ref: schemas/ingestRecordRequest.json
      responses:
        "200":
          description: Records accepted.
          content:
            application/json:
              schema:
                $ref: schemas/ingestRecordResponse.json
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
    delete:
      description: Delete global records.
      operationId: deleteGlobalRecords
      parameters:
        - $ref: parameters/query.yaml
      responses:
        "204":
          description: Records deleted
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
  /meta-storage/records/{globalId}:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: globalId
        required: true
        description: global identifier
        schema:
          type: string
          format: uuid
    get:
      description: Get record with global identifier.
      operationId: getGlobalRecord
      responses:
        "200":
          description: global record found.
          content:
            application/json:
              schema:
                $ref: schemas/globalRecord.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  /meta-storage/clusters:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - $ref: parameters/limit.yaml
      - $ref: parameters/matchkeyid.yaml
      - $ref: parameters/query.yaml
      - $ref: parameters/offset.yaml
    get:
      description: Get clusters with matchkeyid. CQL query with matchValue, clusterId fields
      operationId: getClusters
      responses:
        "200":
          description: cluster record(s) found
          content:
            application/json:
              schema:
                $ref: schemas/clusters.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  /meta-storage/clusters/{clusterId}:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: clusterId
        required: true
        description: cluster identifier
        schema:
          type: string
          format: uuid
    get:
      description: Get cluster by identifier
      operationId: getCluster
      responses:
        "200":
          description: cluster found
          content:
            application/json:
              schema:
                $ref: schemas/cluster.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  /meta-storage/oai:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - $ref: parameters/oai-identifier.yaml
      - $ref: parameters/oai-from.yaml
      - $ref: parameters/oai-list-limit.yaml
      - $ref: parameters/oai-until.yaml
      - $ref: parameters/oai-resumptionToken.yaml
      - $ref: parameters/oai-set.yaml
      - $ref: parameters/oai-verb.yaml
      - $ref: parameters/oai-metadataPrefix.yaml
    get:
      description: OAI service
      operationId: oaiService
      responses:
        "200":
          description: OAI response (possibly with OAI error code)
          content:
            text/xml:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
  /meta-storage/pmh-clients:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
    post:
      description: Create OAI PMH client job
      operationId: postOaiPmhClient
      requestBody:
        content:
          application/json:
            schema:
              $ref: schemas/oai-pmh-client.json
      responses:
        "201":
          description: OAI-PMH client job created
          content:
            application/json:
              schema:
                $ref: schemas/oai-pmh-client.json
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
    get:
      description: Get all OAI PMH client jobs
      operationId: getCollectionOaiPmhClient
      responses:
        "201":
          description: OAI-PMH client job created
          content:
            application/json:
              schema:
                $ref: schemas/oai-pmh-client.json
        "400":
          $ref: "#/components/responses/trait_400"
        "500":
          $ref: "#/components/responses/trait_500"
  /meta-storage/pmh-clients/{id}:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: id
        required: true
        description: OAI-PMH client identifier
        schema:
          type: string
    get:
      description: Get OAI-PMH client
      operationId: getOaiPmhClient
      responses:
        "200":
          description: OAI-PMH client returned
          content:
            application/json:
              schema:
                $ref: schemas/oai-pmh-client.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    put:
      description: Update OAI-PMH client
      operationId: putOaiPmhClient
      requestBody:
        content:
          application/json:
            schema:
              $ref: schemas/oai-pmh-client.json
      responses:
        "204":
          description: updated OK
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
    delete:
      description: Delete OAI-PMH client
      operationId: deleteOaiPmhClient
      responses:
        "200":
          description: OAI-PMH client returned
          content:
            application/json:
              schema:
                $ref: schemas/oai-pmh-client.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  /meta-storage/pmh-clients/{id}/start:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: id
        required: true
        description: OAI-PMH client identifier
        schema:
          type: string
    post:
      description: Start OAI PMH client job
      operationId: startOaiPmhClient
      responses:
        "204":
          description: started ok
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  /meta-storage/pmh-clients/{id}/stop:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: id
        required: true
        description: OAI-PMH client identifier
        schema:
          type: string
    post:
      description: Stop OAI PMH client job
      operationId: stopOaiPmhClient
      responses:
        "204":
          description: stopped ok
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
  /meta-storage/pmh-clients/{id}/status:
    parameters:
      - $ref: headers/okapi-tenant.yaml
      - $ref: headers/okapi-token.yaml
      - $ref: headers/okapi-url.yaml
      - in: path
        name: id
        required: true
        description: OAI-PMH client identifier
        schema:
          type: string
    get:
      description: Get OAI PMH client status
      operationId: statusOaiPmhClient
      responses:
        "200":
          description: status
          content:
            application/json:
              schema:
                $ref: schemas/oai-pmh-client-status.json
        "400":
          $ref: "#/components/responses/trait_400"
        "404":
          $ref: "#/components/responses/trait_404"
        "500":
          $ref: "#/components/responses/trait_500"
# Meta storage error responses
components:
  responses:
    trait_400:
      description: Bad request
      content:
        text/plain:
          schema:
            type: string
            example: Invalid JSON in request
        application/json:
          schema:
            type: object
            example: {"error":"Invalid JSON in request"}
    trait_404:
      description: Not Found
      content:
        text/plain:
          schema:
            type: string
            example: Identifier 596d9f60-cda3-44d2-a4a1-2f48b7d4d23c not found
        application/json:
          schema:
            type: object
            example: {"error":"Identifier 596d9f60-cda3-44d2-a4a1-2f48b7d4d23c not found"}
    trait_500:
      description: Internal error
      content:
        text/plain:
          schema:
            type: string
            example: Internal server error, contact administrator
  schemas:
    errors:
      $ref: schemas/errors.json